requirements.txt
nest-asyncio
aiogram
python-dotenv
bscscan-python
random-text-generator